refactor(client): 抽取 TabContent 单元格与 + 菜单构建到 sidebar/TabContent.tsx
Sidebar.tsx 拆分第一块：memo 化的 TabContent（含 TabContentProps 与
tabContentCompare 比较器）和 buildNewTabOptions 原样迁出到
src/client/sidebar/TabContent.tsx，注释逐字保留。

行为不变：组件体零改动，仅模块路径与 import 调整（相对路径加一层
../）；Sidebar.tsx 改为 import 该模块，对外导出面不变（仍只有
Sidebar 具名导出）。
